requirements.txt
numpy>=1.21.1
Pillow>=9.0.1
scikit_image>=0.19.1
